# fix multiple zeros on split_thousand