--- a/src/lang/en.rs
+++ b/src/lang/en.rs
@@ -76,7 +76,7 @@ impl English {
 
         while !num.is_zero() {
             thousands.push((num % bf_1000).to_u64().unwrap());
-            num /= bf_1000;
+            num = num.div(&bf_1000).int();
         }
 
         thousands