--- a/src/lang/uk.rs
+++ b/src/lang/uk.rs
@@ -663,7 +663,7 @@ impl Ukrainian {
 
         while !num.is_zero() {
             thousands.push((num % bf_1000).to_u64().unwrap());
-            num /= bf_1000;
+            num = num.div(&bf_1000).int();
         }
 
         thousands